Ported many functions from Lua 5.3 to Lua 5.1 (JIT)
The goal is to make Lua 5.1 files work via JIT instead of being
interpreted. Each missing function is replaced by a macro so that
the arguments stay untouched:

ffi::lua_addfunction(a, b)

becomes:

lua_addfunction!(a, b)

This way you can port the functions back to Lua 5.3 effortlessly
in the future. Currently the latest version of LuaJIT is only 5.1
so that's what I am stuck with. However, I'd take a fast version
of 5.1 over a slow version of 5.3 any day.